+++ dapu.py
@@ -7,15 +7,18 @@ import subprocess
 import shutil
 import os
 import logging
-
+user = str(os.environ.get("SUDO_USER"))
 pm=""
+ext_list=False
 supported_pms=["pacman", "apt", "dnf", "zypper"]
-options=["Install packages", "Update the system", "Remove a package", "List installed packages", "Search for a package in the repos", "Clear the package cache", "Exit"]
+supported_aurhelpers=["yay", "paru"]
+options=["Install packages", "Update the system", "Remove a package", "List installed packages", "Search for a package in the repos", "Clear the package cache", "Advanced options", "Exit"]
+adv_options=["Mark a package"]
 commands={
-        "pacman":{"install":["pacman", "-S", "--noconfirm"], "update":["pacman", "-Syu", "--noconfirm"], "remove":["pacman", "-Rns", "--noconfirm"], "query":["pacman", "-Q"], "search":["pacman", "-Ss"], "cache":["pacman", "-Sc", "--noconfirm"]},
-        "apt-get":{"install":["apt-get", "install", "-y"], "update":["apt-get", "update"], "upgrade":["apt-get", "upgrade", "-y"] ,"remove":["apt-get", "purge", "-y"], "query":["apt", "list", "--installed"], "search": ["apt", "search"], "cache":["apt-get", "autoclean", "-y"]},
-        "dnf":{"install":["dnf", "install", "-y"], "update":["dnf", "upgrade", "-y"], "remove":["dnf", "remove", "-y"], "query":["dnf", "list", "--installed"], "search":["dnf", "search"], "cache":["dnf", "clean", "all"]},
-        "zypper":{"install":["zypper", "install", "-y"], "update":["zypper", "dup", "-y"], "remove":["zypper", "remove", "--clean-deps", "-y"], "query":["zypper", "se", "-si"], "search":["zypper", "search"], "cache":["zypper", "clean",]}
+        "pacman":{"install":["pacman", "-S"], "update":["pacman", "-Syu"], "remove":["pacman", "-Rns"], "query":["pacman", "-Q"], "search":["pacman", "-Ss"], "cache":["pacman", "-Sc"], "mark":["pacman", "-D", "--asexplicit"]},
+        "apt-get":{"install":["apt-get", "install"], "update":["apt-get", "update"], "upgrade":["apt-get", "upgrade"] ,"remove":["apt-get", "purge", "--autoremove"], "query":["apt", "list", "--installed"], "search": ["apt", "search"], "cache":["apt-get", "autoclean"], "rm_unneeded":["apt-get", "autoremove"], "mark":["apt-mark", "manual"], "sys-up":["apt-get", "dist-upgrade"]},
+        "dnf":{"install":["dnf", "install"], "update":["dnf", "upgrade"], "remove":["dnf", "remove"], "query":["dnf", "list", "--installed"], "search":["dnf", "search"], "cache":["dnf", "clean", "all"], "rm_unneeded":["dnf", "autoremove"], "mark":["dnf", "mark", "user"]},
+        "zypper":{"install":["zypper", "install"], "update":["zypper", "dup"], "remove":["zypper", "remove", "--clean-deps"], "query":["zypper", "se", "-si"], "search":["zypper", "search"], "cache":["zypper", "clean",], "mark":["zypper", "in", "-f"]}
         }
 
 def check_euid():
@@ -41,7 +44,7 @@ def check_pm():
         subprocess.run(["clear"])
         exit(f"Your package manager isn't supported. The supported package managers are: {supported_pms}")
     logging.info(f" Detected package manager: {pm}")
-
+   
 def check_opensuseleap():
     global leap
     try:
@@ -54,7 +57,142 @@ def list_options():
     for i, item in enumerate(options):
         print(i+1, ")", item)
     print("=======================================")
-def check_command():
+def check_aur():
+    global aurhelper
+    if shutil.which("paru"):
+        aurhelper="paru"
+    elif shutil.which("yay"):
+        aurhelper="yay"
+    else:
+        aurhelper="none"
+        setup_aur=input("No AUR helper was found would you like to set one up? [y/n] ")
+        if setup_aur == "y":
+            subprocess.run(["sudo", "-u", user, "/usr/local/bin/dapu.d/aur-setup.sh"])
+            logging.info("Ran bash script to install AUR helper.")
+            if shutil.which("paru"):
+                aurhelper="paru"
+            elif shutil.which("yay"):
+                aurhelper="yay"
+    logging.info(f"Detected AUR helper: {aurhelper}")
+ 
+def pm_specific_adv_options():
+    global ext_list
+    if not ext_list:
+        if pm != "zypper":
+            adv_options.append("Remove unneeded packages")
+        if pm == "pacman":
+            adv_options.extend(["Install from the AUR", "Update AUR packages", "Search for an AUR package", "Clear system and/or AUR packages"])
+        elif pm=="apt-get":
+            adv_options.extend(["Full system upgrade", "Reconfigure all packages", "Fix broken dependencies"])
+        elif pm=="dnf":
+            adv_options.extend(["Full system upgrade", "Minimal update", "Downgrade a package"])
+        elif pm=="zypper":
+            if check_opensuseleap() == True:
+                adv_options.append("System upgrade")
+        adv_options.append("Go back")
+        ext_list=True
+
+def list_adv_options():  
+    print("============ADVANCED OPTIONS=============")
+    for i, item in enumerate(adv_options):
+        print(i+1, ")", item)
+    print("=========================================")
+
+def adv_menu():
+    global ext_list
+    pm_specific_adv_options()
+    if pm == "pacman":
+        check_aur()
+    while True:
+        list_adv_options()
+        command=input(f"What do you want to do? (1-{str(len(adv_options))}) ")
+        if command == str(len(adv_options)):
+            break
+        elif command == "1":
+            pkg=input("Enter the name of the package to mark as user-installed: ").lower()
+            confirm=(f"The package '{pkg}' will be marked as user-installed, do you wish to continue? [y/n] ")
+            subprocess.run(commands[pm]["mark"]+[pkg])
+            logging.info(f" Ran: {' '.join(commands[pm]['mark'] + [pkg])}")
+        elif command == "1":
+            confirm=input("All uneeded packages will be removed, do you wish to continue? [y/n] ")
+            if pm == "pacman":
+                unneeded=subprocess.run(["pacman", "-Qtdq"], capture_output=True, text=True)
+                pkg=unneeded.stdout.splitlines()
+                subprocess.run(["pacman", "-Rns"]+pkg)
+            else:
+                subprocess.run(commands[pm]["rm_unneeded"])
+            print("========UNNEEDED PACKAGES REMOVED=========")
+        elif command == "3":
+            if pm == "pacman":
+                pkg=input("Which AUR package would you like to install? ").lower()
+                confirm =input(f"The package '{pkg}' will be installed, do you wish to continue? [y/n] ")
+                if confirm == "y":
+                    subprocess.run(["sudo", "-u", user ,aurhelper, "-S", pkg])
+                    logging.info(f"Ran: {aurhelper} -S {pkg}")
+                    print("=========INSTALLATION COMPLETED=========")
+            elif pm == "apt-get":
+                confirm=input("A full system upgrade will be run, do you wish to continue? [y/n] ")
+                if confirm =="y":
+                    subprocess.run(commands[pm]["update"])
+                    logging.info(f" Ran: {' '.join(commands[pm]['update'])}")
+                    subprocess.run(commands[pm]["sys-up"])
+                    logging.info(f" Ran: {' '.join(commands[pm]['sys-up'])}")
+                    print("=========SYSTEM UPDATE COMPLETED=========")
+            elif pm == "dnf":
+                confirm=input("The integrity of all install packages will be checked, do you wish to continue? [y/n] ")
+                if confirm == "y":
+                    subprocess.run(["dnf", "check"])
+                    print("=========SYSTEM UPDATE COMPLETED=========")
+            elif pm == "zypper" and check_opensuseleap():
+                confirm=input("A full system upgrade will be run, do you wish to continue? [y/n] ")
+                if confirm == "y":
+                    subprocess.run(commands[pm]["update"])
+                    logging.info(f" Ran: {' '.join(commands[pm]['upgrade'])}")
+        elif command == "4":
+            if pm == "pacman":
+                confirm=input("The system and AUR packages will be updated, continue? [y/n] ")
+                if confirm == "y":
+                    print("Updating...")
+                    subprocess.run(["sudo", "-u", user, aurhelper])
+                    logging.info(f"Ran: {aurhelper}")
+                    print("============UPDATE COMPLETED============")
+            elif pm == "apt-get":
+                confirm=input("All packages will be reconfigured, do you wish to continue? [y/n] ")
+                if confirm == "y":
+                    subprocess.run(["dpkg --configure", "-a"])
+                    logging.info("Ran: dpkg --configure -a")
+            elif pm == "dnf":
+                confirm=input("A minimal update will be run, do you wish to continue? [y/n] ")
+                if confirm == "y":
+                    subprocess.run(["dnf", "upgrade-minimal"])
+                    print("========MINIMAL UPDATE COMPLETED========")
+        elif command == "5":
+            if pm =="pacman":
+                pkg=input("Which package would you like to search for? ").lower()
+                print("==================RESULTS================")
+                subprocess.run(["sudo", "-u", user, aurhelper, "-Ss", pkg])
+                logging.info(f"Ran: {aurhelper} -Ss {pkg}")
+            elif pm == "apt-get":
+                confirm=input("APT will try to fix broken dependencies, do you wish to continue? [y/n] ")
+                if confirm == "y":
+                    subprocess.run(["apt-get", "--fix-broken", "install"])
+                    logging.info("Ran: apt-get --fix-broken install")
+            elif pm == "dnf":
+                pkg=input("Enter the name of the package you want to downgrade ")
+                confirm=input(f"The package '{pkg}' will be downgraded, do you wish to continue? [y/n] ")
+                if confirm=="y":
+                    subprocess.run(["dnf", "downgrade", pkg])
+                    logging.info(f"Ran: dnf downgrade {pkg} ")
+                    print("===========DOWNGRADE COMPLETED===========")
+        elif command == "6":
+            if pm=="pacman":
+                confirm=input("The system and/or AUR packages' cache will be cleaned, do you wish to continue? [y/n] ")
+                subprocess.run([aurhelper, "-Sc"])
+                logging.info(f"Ran {aurhelper} -Sc")
+
+        else:
+            print("Invalid Option")
+def main():
     command=input(f"What do you want to do? (1-{str(len(options))}) ")
     print("=======================================") 
     if command=="1":
@@ -70,18 +208,21 @@ def check_command():
     elif command=="2":
         confirm = input("The system will be updated, do you wish to continue? [y/n] ")
         if confirm == "y":
+            print("Updating the system...")
             if pm=="zypper":
                 check_opensuseleap()
                 if leap:
                     subprocess.run(["zypper", "up", "-y"])
                     logging.info(" Ran: zypper up -y")
-                    return
-            print("Updating the system...")
-            subprocess.run(commands[pm]["update"])
-            logging.info(f" Ran: {' '.join(commands[pm]['update'])}")
-            if pm == "apt-get":
-                subprocess.run(commands[pm]["upgrade"])
-                logging.info(f" Ran: {' '.join(commands[pm]['upgrade'])}")
+                    print("========UPDATE COMPLETED========")
+                else:
+                    pass
+            else:
+                subprocess.run(commands[pm]["update"])
+                logging.info(f" Ran: {' '.join(commands[pm]['update'])}")
+                if pm == "apt-get":
+                    subprocess.run(commands[pm]["upgrade"])
+                    logging.info(f" Ran: {' '.join(commands[pm]['upgrade'])}")
             print("========UPDATE COMPLETED========")
         else:
             pass
@@ -120,6 +261,8 @@ def check_command():
             print("==============CACHE CLEANED==============")
         else:
             pass
+    elif command == "7":
+        adv_menu()
     elif command ==(str(len(options))):
         logging.info(f"{(' Exited.')}")
         exit("Exited successfully")
@@ -132,4 +275,4 @@ init_logs()
 check_pm()
 while True:
     list_options()
-    check_command()
+    main()